+++ index.html
@@ -4,6 +4,8 @@
   <meta charset="utf-8">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="vendor/leaflet/leaflet.css">
+  <link rel="stylesheet" href="vendor/nouislider/nouislider.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 </head>
@@ -217,7 +219,7 @@
       </ul>
       <p class="popup__description">Великолепная квартира-студия в центре Токио. Подходит как туристам, так и бизнесменам. Квартира полностью укомплектована и недавно отремонтирована.</p>
       <div class="popup__photos">
-        <img src="" class="popup__photo" width="45" height="40" alt="Фотография жилья">
+        <img src="#" class="popup__photo" width="45" height="40" alt="Фотография жилья">
       </div>
     </article>
   </template>
@@ -237,6 +239,8 @@
     </div>
   </template>
   <script src="vendor/pristine/pristine.min.js" type="text/javascript"></script>
+  <script src="vendor/nouislider/nouislider.js"></script>
+  <script src="vendor/leaflet/leaflet.js"></script>
   <script src="js/main.js" type="module"></script>
 </body>
 </html>
