+++ js/form.js
@@ -1,3 +1,5 @@
+import { ROOMS_OPTION, ROOMS_ERRORS, TYPE_OPTINS } from './loading.js';
+
 const form = document.querySelector('.ad-form');
 const pristine = new Pristine(form, {
   classTo: 'ad-form__element',
@@ -5,51 +7,33 @@ const pristine = new Pristine(form, {
   errorTextParent: 'ad-form__element',
   errorTextClass: 'text-help'
 });
-form.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  pristine.validate();
-});
+
 const roomsField = form.querySelector('#room_number');
 const capacityField = form.querySelector('#capacity');
-const roomErrors = {
-  1: 'Только для 1 гостя',
-  2: 'Для 2 гостей и меньше',
-  3: 'Для 3 гостей и меньше',
-  100: 'Не для гостей'
-};
-const roomsOption = {
-  1: ['1'],
-  2: ['1', '2'],
-  3: ['1', '2', '3'],
-  100: ['0']
-};
-const getRoomsErrorMessage = () => roomErrors[roomsField.value];
-const validateRooms = () => roomsOption[roomsField.value].includes(capacityField.value);
-pristine.addValidator(roomsField, validateRooms, getRoomsErrorMessage);
-pristine.addValidator(capacityField, validateRooms);
-roomsField.addEventListener('change', () => pristine.validate(capacityField));
-capacityField.addEventListener('change', () => pristine.validate(roomsField));
-
 const typeField = form.querySelector('#type');
 const priceField = form.querySelector('#price');
 const timeinField = form.querySelector('#timein');
 const timeoutField = form.querySelector('#timeout');
 
-const typeOption = {
-  bungalow: 0,
-  flat: 1000,
-  hotel: 3000,
-  house: 5000,
-  palace: 10000
-};
-
-typeField.addEventListener('change', () => {
-  priceField.placeholder = typeOption[typeField.value];
+form.addEventListener('submit', (evt) => {
+  evt.preventDefault();
+  pristine.validate();
 });
 
-const validatePrice = () => priceField.value >= typeOption[typeField.value];
+const getRoomsErrorMessage = () => ROOMS_ERRORS[roomsField.value];
+const validateRooms = () => ROOMS_OPTION[roomsField.value].includes(capacityField.value);
+pristine.addValidator(roomsField, validateRooms, getRoomsErrorMessage);
+pristine.addValidator(capacityField, validateRooms);
+roomsField.addEventListener('change', () => pristine.validate(capacityField));
+capacityField.addEventListener('change', () => pristine.validate(roomsField));
 
-const getPriceErrorMessage = () => `Минимальная цена ${typeOption[typeField.value]}`;
+typeField.addEventListener('change', () => {
+  priceField.placeholder = TYPE_OPTINS[typeField.value];
+});
+
+const validatePrice = () => priceField.value >= TYPE_OPTINS[typeField.value];
+
+const getPriceErrorMessage = () => `Минимальная цена ${TYPE_OPTINS[typeField.value]}`;
 
 pristine.addValidator(priceField, validatePrice, getPriceErrorMessage);
 
@@ -72,6 +56,7 @@ const switchStateForm = (state) => {
   form.classList.toggle('ad-form--disabled', state);
   switchStateElements(fieldsets, state);
 };
+
 const switchStateFilter = (state) => {
   const filter = document.querySelector('.map__filters');
   const selects = filter.querySelectorAll('select');
@@ -80,10 +65,14 @@ const switchStateFilter = (state) => {
   switchStateElements(selects, state);
   switchStateElements(fieldsets, state);
 };
+
 const switchStatePage = (state) => {
   switchStateForm(state);
   switchStateFilter(state);
 };
+
 const deactivatePage = () => switchStatePage(true);
+
 const activatePage = () => switchStatePage(false);
+
 export { deactivatePage, activatePage };
