+++ js/loading.js
@@ -64,6 +64,33 @@ export const OFFER_FEATURES = {
   conditioner: 'Кондиционер'
 };
 
+export const ROOMS_ERRORS = {
+  1: 'Только для 1 гостя',
+  2: 'Для 2 гостей и меньше',
+  3: 'Для 3 гостей и меньше',
+  100: 'Не для гостей'
+};
+
+export const ROOMS_OPTION = {
+  1: ['1'],
+  2: ['1', '2'],
+  3: ['1', '2', '3'],
+  100: ['0']
+};
+
+export const TYPE_OPTINS = {
+  bungalow: 0,
+  flat: 1000,
+  hotel: 3000,
+  house: 5000,
+  palace: 10000
+};
+
 export const ELEMENT_COUNT = 10;
 
-export default {TITLES, TYPES, TIMES, FEATURES, DESCRIPTIONS, PHOTOS, OFFER_TYPE, OFFER_FEATURES, ELEMENT_COUNT};
+export const DEFAULT_COORDINATE = {
+  lat: 35.6895000,
+  lng: 139.6917100,
+};
+
+export default {TITLES, TYPES, TIMES, FEATURES, DESCRIPTIONS, PHOTOS, OFFER_TYPE, OFFER_FEATURES, ELEMENT_COUNT, ROOMS_OPTION, ROOMS_ERRORS, TYPE_OPTINS, DEFAULT_COORDINATE};
